move request class into action class, move all classes from xpack to server module, consistent naming
