+++ server/src/main/java/org/elasticsearch/action/datastreams/ModifyDataStreamsTransportAction.java
@@ -1,11 +1,12 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
-package org.elasticsearch.xpack.datastreams.action;
+package org.elasticsearch.action.datastreams;
 
 import org.elasticsearch.action.ActionListener;
 import org.elasticsearch.action.support.ActionFilters;
@@ -21,15 +22,14 @@ import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.tasks.Task;
 import org.elasticsearch.threadpool.ThreadPool;
 import org.elasticsearch.transport.TransportService;
-import org.elasticsearch.xpack.core.action.ModifyDataStreamsAction;
 
-public class ModifyDataStreamTransportAction extends AcknowledgedTransportMasterNodeAction<
-    MetadataDataStreamsService.ModifyDataStreamRequest> {
+public class ModifyDataStreamsTransportAction extends AcknowledgedTransportMasterNodeAction<
+    ModifyDataStreamsAction.Request> {
 
     private final MetadataDataStreamsService metadataDataStreamsService;
 
     @Inject
-    public ModifyDataStreamTransportAction(
+    public ModifyDataStreamsTransportAction(
         TransportService transportService,
         ClusterService clusterService,
         ThreadPool threadPool,
@@ -43,7 +43,7 @@ public class ModifyDataStreamTransportAction extends AcknowledgedTransportMaster
             clusterService,
             threadPool,
             actionFilters,
-            MetadataDataStreamsService.ModifyDataStreamRequest::new,
+            ModifyDataStreamsAction.Request::new,
             indexNameExpressionResolver,
             ThreadPool.Names.SAME
         );
@@ -53,7 +53,7 @@ public class ModifyDataStreamTransportAction extends AcknowledgedTransportMaster
     @Override
     protected void masterOperation(
         Task task,
-        MetadataDataStreamsService.ModifyDataStreamRequest request,
+        ModifyDataStreamsAction.Request request,
         ClusterState state,
         ActionListener<AcknowledgedResponse> listener
     ) throws Exception {
@@ -61,7 +61,7 @@ public class ModifyDataStreamTransportAction extends AcknowledgedTransportMaster
     }
 
     @Override
-    protected ClusterBlockException checkBlock(MetadataDataStreamsService.ModifyDataStreamRequest request, ClusterState state) {
+    protected ClusterBlockException checkBlock(ModifyDataStreamsAction.Request request, ClusterState state) {
         return state.blocks().globalBlockedException(ClusterBlockLevel.METADATA_WRITE);
     }
