+++ server/src/main/java/org/elasticsearch/cluster/metadata/MetadataDataStreamsService.java
@@ -9,30 +9,18 @@
 package org.elasticsearch.cluster.metadata;
 
 import org.elasticsearch.action.ActionListener;
-import org.elasticsearch.action.ActionRequestValidationException;
-import org.elasticsearch.action.IndicesRequest;
-import org.elasticsearch.action.support.IndicesOptions;
-import org.elasticsearch.action.support.master.AcknowledgedRequest;
+import org.elasticsearch.action.datastreams.ModifyDataStreamsAction;
 import org.elasticsearch.action.support.master.AcknowledgedResponse;
 import org.elasticsearch.cluster.AckedClusterStateUpdateTask;
 import org.elasticsearch.cluster.ClusterState;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.Priority;
-import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.index.mapper.MapperService;
 import org.elasticsearch.indices.IndicesService;
-import org.elasticsearch.xcontent.ObjectParser;
-import org.elasticsearch.xcontent.ParseField;
-import org.elasticsearch.xcontent.ToXContentObject;
-import org.elasticsearch.xcontent.XContentBuilder;
 
 import java.io.IOException;
-import java.util.Collections;
-import java.util.List;
 import java.util.function.Function;
-
-import static org.elasticsearch.action.ValidateActions.addValidationError;
 
 /**
  * Handles data stream modification requests.
@@ -47,7 +35,7 @@ public class MetadataDataStreamsService {
         this.indicesService = indicesService;
     }
 
-    public void modifyDataStream(final ModifyDataStreamRequest request,
+    public void modifyDataStream(final ModifyDataStreamsAction.Request request,
                                  final ActionListener<AcknowledgedResponse> listener) {
         if (request.getActions().size() == 0) {
             listener.onResponse(AcknowledgedResponse.TRUE);
@@ -165,72 +153,4 @@ public class MetadataDataStreamsService {
         return index;
     }
 
-    public static final class ModifyDataStreamRequest
-        extends AcknowledgedRequest<ModifyDataStreamRequest>
-        implements IndicesRequest, ToXContentObject {
-
-        // relevant only for authorizing the request, so require every specified
-        // index to exist, expand wildcards only to open indices, prohibit
-        // wildcard expressions that resolve to zero indices, and do not attempt
-        // to resolve expressions as aliases
-        private static final IndicesOptions INDICES_OPTIONS =
-            IndicesOptions.fromOptions(false, false, true, false, true, false, true, false);
-
-        private List<DataStreamAction> actions;
-
-        public ModifyDataStreamRequest() {}
-
-        public ModifyDataStreamRequest(StreamInput in) throws IOException {
-            super(in);
-            actions = in.readList(DataStreamAction::new);
-        }
-
-        public ModifyDataStreamRequest(List<DataStreamAction> actions) {
-            this.actions = Collections.unmodifiableList(actions);
-        }
-
-        public List<DataStreamAction> getActions() {
-            return actions;
-        }
-
-        public void setActions(List<DataStreamAction> actions) {
-            this.actions = Collections.unmodifiableList(actions);
-        }
-
-        @Override
-        public XContentBuilder toXContent(XContentBuilder builder, Params params) throws IOException {
-            builder.startObject();
-            builder.startArray("actions");
-            for (DataStreamAction action : actions) {
-                action.toXContent(builder, params);
-            }
-            builder.endArray();
-            builder.endObject();
-            return builder;
-        }
-
-        @Override
-        public ActionRequestValidationException validate() {
-            if (actions.isEmpty()) {
-                return addValidationError("must specify at least one data stream modification action", null);
-            }
-            return null;
-        }
-
-        public static final ObjectParser<ModifyDataStreamRequest, Void> PARSER =
-            new ObjectParser<>("data_stream_actions", ModifyDataStreamRequest::new);
-        static {
-            PARSER.declareObjectArray(ModifyDataStreamRequest::setActions, DataStreamAction.PARSER, new ParseField("actions"));
-        }
-
-        @Override
-        public String[] indices() {
-            return actions.stream().map(DataStreamAction::getDataStream).toArray(String[]::new);
-        }
-
-        @Override
-        public IndicesOptions indicesOptions() {
-            return INDICES_OPTIONS;
-        }
-    }
 }
