--- x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/action/ModifyDataStreamsAction.java
@@ -1,21 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-package org.elasticsearch.xpack.core.action;
-
-import org.elasticsearch.action.ActionType;
-import org.elasticsearch.action.support.master.AcknowledgedResponse;
-
-public class ModifyDataStreamsAction extends ActionType<AcknowledgedResponse> {
-
-    public static final ModifyDataStreamsAction INSTANCE = new ModifyDataStreamsAction();
-    public static final String NAME = "indices:admin/data_stream/modify";
-
-    private ModifyDataStreamsAction() {
-        super(NAME, AcknowledgedResponse::readFrom);
-    }
-}
